Reorganizes project files to make it easier to read.
